Refactor: useItemQuery 파일에서 mutation 에러 핸들하기
Related to: #377
