+++ src/components/Main/MyChallenge/MyChallengeModal/GetRewardTwiceModal/GetRewardTwiceModal.tsx
@@ -1,12 +1,10 @@
 import LoadingBox from "@/components/Common/Loading/LoadingBox/LoadingBox";
 import { Modal } from "@/components/Common/Modal/Modal";
-import { usePostPointTwiceItemUse } from "@/hooks/queries/useItemQuery";
 import { useEffect } from "react";
 import { MyChallengeDoneDataType } from "@/types/myChallengeType";
 import CommonModal from "@/components/Common/CommonModal/CommonModal";
-import CommonMutationErrorModal from "@/components/Error/CommonMutationErrorModal/CommonMutationErrorModal";
-import { AxiosError } from "axios";
 import { useModalStore } from "@/stores/modalStore";
+import { usePostPointTwiceItemUse } from "@/hooks/queries/useItemQuery";
 
 interface Props {
   instanceId: number;
@@ -25,18 +23,11 @@ function GetRewardTwiceModal({ instanceId, itemId }: Props) {
     );
   };
 
-  const onErrorPostItemUse = (error: AxiosError<{ message?: string }>) => {
-    setModal(
-      <CommonMutationErrorModal error={error} closeModal={closeModal} />
-    );
-  };
-
   const {
     mutate: pointTwiceItemUseMutate,
     isLoading: pointTwiceItemUseLoading,
   } = usePostPointTwiceItemUse({
     onSuccess: onSuccessPostItemUse,
-    onError: onErrorPostItemUse,
   });
 
   useEffect(() => {
