+++ src/hooks/queries/useItemQuery.tsx
@@ -12,15 +12,19 @@ import getItemPassApi from "@/apis/getItemPassApi";
 import getItemPointApi from "@/apis/getItemPointApi";
 import { MyChallengeDoneDataType } from "@/types/myChallengeType";
 import { AxiosError } from "axios";
+import { useModalStore } from "@/stores/modalStore";
+import CommonMutationErrorModal from "@/components/Error/CommonMutationErrorModal/CommonMutationErrorModal";
+import { useNavigate } from "react-router-dom";
+import { PATH } from "@/constants/path";
+import CommonModal from "@/components/Common/CommonModal/CommonModal";
 
 interface PostFrameItemEquiptmentType {
-  onError: (error: AxiosError<{ message?: string }>) => void;
-  onSuccess?: () => void;
+  onSuccess: () => void;
 }
 export const usePostFrameItemEquiptment = ({
-  onError,
   onSuccess,
 }: PostFrameItemEquiptmentType) => {
+  const { setModal, closeModal } = useModalStore();
   const queryClient = useQueryClient();
   const { mutate } = useMutation(
     (itemId: number) => postItemEquipApi({ itemId }),
@@ -28,21 +32,20 @@ export const usePostFrameItemEquiptment = ({
       onSuccess: (data) => {
         localStorage.setItem(FRAMEID, encrypt(data.itemId));
         queryClient.invalidateQueries(QUERY_KEY.SHOP_FRAME_ITEMS);
-        onSuccess && onSuccess();
+        onSuccess();
       },
-      onError: (error: AxiosError<{ message?: string }>) => onError(error),
+      onError: (error: AxiosError<{ message?: string }>) => {
+        setModal(
+          <CommonMutationErrorModal error={error} closeModal={closeModal} />
+        );
+      },
     }
   );
   return { mutate };
 };
 
-interface PostFrameItemUnEquiptmentType {
-  onError: (error: AxiosError<{ message?: string }>) => void;
-}
-
-export const usePostFrameItemUnEquiptment = ({
-  onError,
-}: PostFrameItemUnEquiptmentType) => {
+export const usePostFrameItemUnEquiptment = () => {
+  const { setModal, closeModal } = useModalStore();
   const queryClient = useQueryClient();
   const { mutate, mutateAsync } = useMutation(postItemUnEquipApi, {
     onSuccess: () => {
@@ -50,7 +53,9 @@ export const usePostFrameItemUnEquiptment = ({
       queryClient.invalidateQueries(QUERY_KEY.SHOP_FRAME_ITEMS);
     },
     onError: (error: AxiosError<{ message?: string }>) => {
-      onError(error);
+      setModal(
+        <CommonMutationErrorModal error={error} closeModal={closeModal} />
+      );
     },
   });
   return { mutate, mutateAsync };
@@ -63,13 +68,12 @@ interface usePostItemUseMutateType {
 
 interface usePostItemUseType {
   onSuccess: () => void;
-  onError: (error: AxiosError<{ message?: string }>) => void;
 }
 
 export const usePostCertificationPassItemUse = ({
   onSuccess,
-  onError,
 }: usePostItemUseType) => {
+  const { setModal, closeModal } = useModalStore();
   const queryClient = useQueryClient();
   const { mutate, isLoading } = useMutation(
     ({ instanceId, itemId }: usePostItemUseMutateType) =>
@@ -79,7 +83,11 @@ export const usePostCertificationPassItemUse = ({
         queryClient.invalidateQueries(QUERY_KEY.MY_ACTIVITY_CHALLENGES);
         onSuccess();
       },
-      onError: (error: AxiosError<{ message?: string }>) => onError(error),
+      onError: (error: AxiosError<{ message?: string }>) => {
+        setModal(
+          <CommonMutationErrorModal error={error} closeModal={closeModal} />
+        );
+      },
     }
   );
   return { mutate, isLoading };
@@ -87,12 +95,11 @@ export const usePostCertificationPassItemUse = ({
 
 interface PostPointTwiceItemUseType {
   onSuccess: (res: MyChallengeDoneDataType) => void;
-  onError: (error: AxiosError<{ message?: string }>) => void;
 }
 export const usePostPointTwiceItemUse = ({
   onSuccess,
-  onError,
 }: PostPointTwiceItemUseType) => {
+  const { setModal, closeModal } = useModalStore();
   const queryClient = useQueryClient();
   const { mutate, isLoading } = useMutation(
     ({ instanceId, itemId }: usePostItemUseMutateType) =>
@@ -103,7 +110,9 @@ export const usePostPointTwiceItemUse = ({
         onSuccess(res);
       },
       onError: (error: AxiosError<{ message?: string }>) => {
-        onError(error);
+        setModal(
+          <CommonMutationErrorModal error={error} closeModal={closeModal} />
+        );
       },
     }
   );
@@ -112,12 +121,12 @@ export const usePostPointTwiceItemUse = ({
 
 interface usePostItemBuyType {
   onSuccess: () => void;
-  onError: (error: AxiosError<{ message?: string }>) => void;
 }
 export const usePostItemBuy = ({
   onSuccess: onSuccess,
-  onError: onError,
 }: usePostItemBuyType) => {
+  const { setModal, closeModal } = useModalStore();
+  const navigate = useNavigate();
   const queryClient = useQueryClient();
   const { mutate, isLoading } = useMutation(
     (itemId: number) => postdItemBuyApi({ itemId }),
@@ -129,8 +138,40 @@ export const usePostItemBuy = ({
         queryClient.invalidateQueries(QUERY_KEY.MY_PROFILE);
         onSuccess();
       },
+
       onError: (error: AxiosError<{ message?: string }>) => {
-        onError(error);
+        const onClickMoveToCharge = () => {
+          closeModal();
+          navigate(PATH.PAYMENTS);
+        };
+
+        if (
+          error?.response?.data.message ===
+          "프로필 프레임은 재구매가 불가능 합니다."
+        )
+          setModal(
+            <CommonModal
+              content={"이미 소지하고 있는 아이템입니다!"}
+              buttonContent="확인"
+              onClick={closeModal}
+            />
+          );
+        else if (
+          error?.response?.data.message ===
+          "사용자의 보유 포인트가 충분하지 않습니다."
+        ) {
+          setModal(
+            <CommonModal
+              content={"포인트 잔액이 부족해요\n충전하시겠어요?"}
+              buttonContent="충전하러 가기"
+              onClick={onClickMoveToCharge}
+            />
+          );
+        } else {
+          setModal(
+            <CommonMutationErrorModal error={error} closeModal={closeModal} />
+          );
+        }
       },
     }
   );
